correction des axes 2Dto3D

+++ src/MASTER/labyrinthe.hpp
@@ -3,8 +3,8 @@
 
 class labyrinthe{
     public:
-    cv::Point2f PointDepart;
-    cv::Point2f PointArrivee;
+    cv::Point2d PointDepart;
+    cv::Point2d PointArrivee;
     std::vector<cv::Vec4f> lignes;
 };
